refactor: split command palette components
